+++ changelog.md
@@ -8,6 +8,8 @@ Codion Change Log
 - TableConditionModel, add/removeListener methods replaced with methods that return the underlying event observer.
 - FilteredModel.Refresher, add/removeListener methods replaced with methods that return the underlying event observer.
 - LoadTest, add/removeListener methods replaced with methods that return the underlying event observer.
+### is.codion.framework.db
+- EntityConnectionProvider, add/removeListener methods replaced with methods that return the underlying event observer.
 ### is.codion.framework.model
 - EntityEditModel.setDefaults() renamed defaults().
 - EntityEditModel.setDefault() renamed defaultValue(), now Value based.